requirements.txt
google_api_python_client==2.78.0
hikari==2.0.0.dev121
hikari_lightbulb==2.3.3
python-dotenv==0.21.1
requests==2.27.1
colorlog==6.7.0
APScheduler==3.10.1
hikari-miru==3.1.1
git+https://github.com/bachtran02/Lavalink.py.git@custom-branch